requirements.txt
flask==3.0.0
pinecone-client==3.0.0
openai==1.12.0
gunicorn==21.2.0
